+++ src/base/filter.rs
@@ -2,18 +2,18 @@ use crate::pbrt::*;
 
 pub struct PixelSample {
     pub p: Point2f,
-    pub weight: Float,
+    pub weight: f64,
 }
 
 pub struct FilterSample {
     pub p: Point2f,
-    pub weight: Float,
+    pub weight: f64,
 }
 
 pub trait Filter: Send + Sync {
-    fn get_integral(&self) -> Float;
+    fn get_integral(&self) -> f64;
 
     fn sample(&self, u: Point2f) -> FilterSample;
 
-    fn evaluate(&self, p: Point2f) -> Float;
+    fn evaluate(&self, p: Point2f) -> f64;
 }
