+++ src/euclidean_space/compensated_float.rs
@@ -1,36 +1,36 @@
 use crate::pbrt::*;
 
 pub struct CompensatedFloat {
-    pub value: Float,
-    pub error: Float,
+    pub value: f64,
+    pub error: f64,
 }
 
 impl CompensatedFloat {
-    pub fn new(_value: Float, _error: Float) -> Self {
+    pub fn new(_value: f64, _error: f64) -> Self {
         return CompensatedFloat {
             value: _value,
             error: _error,
         };
     }
 
-    pub fn eval(&self) -> Float {
+    pub fn eval(&self) -> f64 {
         return self.value + self.error;
     }
 }
 
-pub fn two_prod(a: Float, b: Float) -> CompensatedFloat {
+pub fn two_prod(a: f64, b: f64) -> CompensatedFloat {
     let ab = a * b;
     return CompensatedFloat::new(ab, fma(a, b, -ab));
 }
 
-pub fn two_sum(a: Float, b: Float) -> CompensatedFloat {
+pub fn two_sum(a: f64, b: f64) -> CompensatedFloat {
     let s = a + b;
     let delta = s - a;
     return CompensatedFloat::new(s, (a - (s - delta)) + (b - delta));
 }
 
-pub fn inner_product<const N: usize>(left: &[Float; N], right: &[Float; N]) -> CompensatedFloat {
-    fn _inner_product(left: &[Float], right: &[Float], n: usize) -> CompensatedFloat {
+pub fn inner_product<const N: usize>(left: &[f64; N], right: &[f64; N]) -> CompensatedFloat {
+    fn _inner_product(left: &[f64], right: &[f64], n: usize) -> CompensatedFloat {
         let ab = two_prod(left[0], right[0]);
         if n == 1 {
             return ab;
